+++ src/SpeedyWeather.jl
@@ -175,10 +175,10 @@ include("physics/land/land.jl")
 
 # OUTPUT
 include("output/schedule.jl")
+include("output/callbacks.jl")
 include("output/feedback.jl")
 include("output/netcdf_output.jl")
 include("output/restart_file.jl")
-include("output/callbacks.jl")
 include("output/particle_tracker.jl")
 include("output/jld2_output.jl")