+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # Unreleased
 
 - Expand scratch memory handling in spectral transforms for better differentiability with Enzyme and revise FFT plans [#791](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/791)
+- Reimplement RestartFile, ProgressTxt, ParamatersTxt as callback [#726](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/726/)
+- Implicit has reinitialize option [#726](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/726/)
+- Leapfrog restart default [#726](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/726/)
 - Reduce GPU overhead and revise on_architecture and adapt [#824](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/824)
 - Update CITATION.cff for Zenodo [#869](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/869)
 - 10m winds and surface temperature output [#858](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/858)